[Refactor] Refactor cluster Get certs
