+++ openstack/cce/v3/clusters/results.go
@@ -6,63 +6,6 @@ import (
 	golangsdk "github.com/opentelekomcloud/gophertelekomcloud"
 )
 
-type Certificate struct {
-	// API type, fixed value Config
-	Kind string `json:"kind"`
-	// API version, fixed value v1
-	ApiVersion string `json:"apiVersion"`
-	// Cluster list
-	Clusters []CertClusters `json:"clusters"`
-	// User list
-	Users []CertUsers `json:"users"`
-	// Context list
-	Contexts []CertContexts `json:"contexts"`
-	// The current context
-	CurrentContext string `json:"current-context"`
-}
-
-type CertClusters struct {
-	// Cluster name
-	Name string `json:"name"`
-	// Cluster information
-	Cluster CertCluster `json:"cluster"`
-}
-
-type CertCluster struct {
-	// Server IP address
-	Server string `json:"server"`
-	// Certificate data
-	CertAuthorityData string `json:"certificate-authority-data"`
-}
-
-type CertUsers struct {
-	// User name
-	Name string `json:"name"`
-	// Cluster information
-	User CertUser `json:"user"`
-}
-
-type CertUser struct {
-	// Client certificate
-	ClientCertData string `json:"client-certificate-data"`
-	// Client key data
-	ClientKeyData string `json:"client-key-data"`
-}
-
-type CertContexts struct {
-	// Context name
-	Name string `json:"name"`
-	// Context information
-	Context CertContext `json:"context"`
-}
-
-type CertContext struct {
-	// Cluster name
-	Cluster string `json:"cluster"`
-	// User name
-	User string `json:"user"`
-}
-
 // UnmarshalJSON helps to unmarshal Status fields into needed values.
 // OTC and Huawei have different data types and child fields for `endpoints` field in Cluster Status.
 // This function handles the unmarshal for both
@@ -100,12 +43,6 @@ func (r *Status) UnmarshalJSON(b []byte) error {
 	r.Endpoints = s.Endpoints
 
 	return err
-}
-
-// DeleteResult represents the result of a delete operation. Call its ExtractErr
-// method to determine if the request succeeded or failed.
-type DeleteResult struct {
-	golangsdk.ErrResult
 }
 
 type GetCertResult struct {
